[Raffle] v1.7.0 Improve mixins
