+++ raffle/commands/editor/editor.py
@@ -5,30 +5,29 @@ from typing import Union
 import discord
 from redbot.core import commands
 from redbot.core.i18n import Translator
-from redbot.core.utils.chat_formatting import box, inline
+from redbot.core.utils.chat_formatting import box
 from redbot.core.utils.menus import start_adding_reactions
 from redbot.core.utils.predicates import MessagePredicate, ReactionPredicate
 
-from ..mixins.abc import RaffleMixin
-from ..utils.checks import VALID_USER_BADGES
-from ..utils.converters import RaffleFactoryConverter
-from ..utils.enums import RaffleComponents
-from ..utils.exceptions import InvalidArgument, RaffleError
-from ..utils.formatting import cross, tick
-from ..utils.helpers import (
+from ...mixins.abc import RaffleMixin
+from ...mixins.metaclass import MetaClass
+from ...utils.converters import RaffleFactoryConverter
+from ...utils.enums import RaffleComponents
+from ...utils.exceptions import InvalidArgument, RaffleError
+from ...utils.formatting import cross, tick
+from ...utils.helpers import (
     cleanup_code,
     format_traceback,
-    format_underscored_text,
     raffle_safe_member_scanner,
     start_interactive_message_session,
     validator,
 )
-from ..utils.parser import RaffleManager
+from ...utils.parser import RaffleManager
 
 _ = Translator("Raffle", __file__)
 
 
-class EditorCommands(RaffleMixin):
+class EditorCommands(RaffleMixin, metaclass=MetaClass):
     """Mixin for commands under ``[p]raffle edit``."""
 
     @commands.group()
@@ -628,424 +627,3 @@ class EditorCommands(RaffleMixin):
         await ctx.send(update)
 
         await self.clean_guild_raffles(ctx)
-
-    @edit.group()
-    async def prevented(self, ctx):
-        """Manage prevented users in a raffle."""
-        pass
-
-    @prevented.command(name="add")
-    async def prevented_add(self, ctx, raffle: RaffleFactoryConverter, member: discord.Member):
-        """Add a member to the prevented list of a raffle.
-
-        **Arguments:**
-            - `<raffle>` - The name of the raffle.
-            - `<member>` - The member to add to the prevented list.
-        """
-        async with self.config.guild(ctx.guild).raffles() as r:
-
-            raffle_data = r.get(raffle, None)
-
-            prevented = raffle_data.get("prevented_users", [])
-
-            if member.id in prevented:
-                return await ctx.send(_("This user is already prevented in this raffle."))
-
-            if not prevented:
-                raffle_data["prevented_users"] = [member.id]
-            else:
-                prevented.append(member.id)
-
-            await ctx.send(
-                _("{} added to the prevented list for this raffle.".format(member.name))
-            )
-
-        await self.clean_guild_raffles(ctx)
-
-    @prevented.command(name="remove", aliases=["del"])
-    async def prevented_remove(self, ctx, raffle: RaffleFactoryConverter, member: discord.Member):
-        """Remove a member from the prevented list of a raffle.
-
-        **Arguments:**
-            - `<raffle>` - The name of the raffle.
-            - `<member>` - The member to remove from the prevented list.
-        """
-        async with self.config.guild(ctx.guild).raffles() as r:
-
-            raffle_data = r.get(raffle, None)
-
-            prevented = raffle_data.get("prevented_users", [])
-
-            if member.id not in prevented:
-                return await ctx.send(_("This user was not already prevented in this raffle."))
-
-            prevented.remove(member.id)
-            if not prevented:
-                del raffle_data["prevented_users"]
-            await ctx.send(
-                _("{} remove from the prevented list for this raffle.".format(member.name))
-            )
-
-        await self.clean_guild_raffles(ctx)
-
-    @prevented.command(name="clear")
-    async def prevented_clear(self, ctx, raffle: RaffleFactoryConverter):
-        """Clear the prevented list for a raffle.
-
-        **Arguments:**
-            - `<raffle>` - The name of the raffle.
-        """
-        async with self.config.guild(ctx.guild).raffles() as r:
-
-            raffle_data = r.get(raffle, None)
-
-            prevented = raffle_data.get("prevented_users", None)
-
-            if prevented is None:
-                return await ctx.send(_("There are no prevented users."))
-
-            message = _("Are you sure you want to clear the prevented users list for this raffle?")
-            can_react = ctx.channel.permissions_for(ctx.me).add_reactions
-            if not can_react:
-                message += " (yes/no)"
-            message = await ctx.send(message)
-            if can_react:
-                start_adding_reactions(message, ReactionPredicate.YES_OR_NO_EMOJIS)
-                predicate = ReactionPredicate.yes_or_no(message, ctx.author)
-                event_type = "reaction_add"
-            else:
-                predicate = MessagePredicate.yes_or_no(ctx)
-                event_type = "message"
-
-            try:
-                await self.bot.wait_for(event_type, check=predicate, timeout=30)
-            except asyncio.TimeoutError:
-                await ctx.send(_("You took too long to respond."))
-                return
-
-            if predicate.result:
-                del raffle_data["prevented_users"]
-                message = _("Prevented users list cleared for this raffle.")
-                try:
-                    await message.edit(content=message)
-                except discord.NotFound:
-                    await ctx.send(message)
-
-            else:
-                await ctx.send(_("No changes have been made."))
-
-    @edit.group()
-    async def badges(self, ctx):
-        """Manage required badges in a raffle."""
-        pass
-
-    @badges.command(name="add")
-    async def badges_add(self, ctx, raffle: RaffleFactoryConverter, *badges: str):
-        """Add a badge to the required badges list of a raffle.
-
-        **Arguments:**
-            - `<raffle>` - The name of the raffle.
-            - `<badges>` - The badge(s) to add to the required badges list.
-        """
-        async with self.config.guild(ctx.guild).raffles() as r:
-
-            raffle_data = r.get(raffle, None)
-
-            badges_list = raffle_data.get("badges_needed_to_enter", [])
-
-            for badge in badges:
-                if badge not in VALID_USER_BADGES:
-                    return await ctx.send(
-                        _('"{}" was not a recognized Discord badge.'.format(badge))
-                    )
-                if badge in badges_list:
-                    return await ctx.send(
-                        _(
-                            'The "{}" badge is already required in this raffle.'.format(
-                                format_underscored_text(badge)
-                            )
-                        )
-                    )
-
-                if not badges_list:
-                    raffle_data["badges_needed_to_enter"] = list(badges)
-                else:
-                    badges_list.append(badge)
-
-            await ctx.send(
-                _(
-                    "Added the following badges as requirements in this raffle: {}.".format(
-                        ", ".join(inline(format_underscored_text(b)) for b in badges)
-                    )
-                )
-            )
-
-        await self.clean_guild_raffles(ctx)
-
-    @badges.command(name="remove", aliases=["del"])
-    async def badges_remove(self, ctx, raffle: RaffleFactoryConverter, *badges: str):
-        """Remove a badge from the required badges list of a raffle.
-
-        **Arguments:**
-            - `<raffle>` - The name of the raffle.
-            - `<member>` - The badge to remove from the required badges list.
-        """
-        async with self.config.guild(ctx.guild).raffles() as r:
-
-            raffle_data = r.get(raffle, None)
-
-            badges_list = raffle_data.get("badges_needed_to_enter", [])
-
-            for badge in badges:
-                if badge not in VALID_USER_BADGES:
-                    return await ctx.send(
-                        _('"{}" was not a recognized Discord badge.'.format(badge))
-                    )
-                if badge not in badges_list:
-                    return await ctx.send(
-                        _('The "{}" badge was not already required in this raffle.'.format(badge))
-                    )
-
-            badges_list.remove(badge)
-            if not badges_list:
-                del raffle_data["badges_list"]
-            await ctx.send(
-                _(
-                    "Added the following badges as requirements in this raffle: {}.".format(
-                        ", ".join(inline(format_underscored_text(b)) for b in badges)
-                    )
-                )
-            )
-
-        await self.clean_guild_raffles(ctx)
-
-    @badges.command(name="clear")
-    async def badges_clear(self, ctx, raffle: RaffleFactoryConverter):
-        """Clear the required badges list for a raffle.
-
-        **Arguments:**
-            - `<raffle>` - The name of the raffle.
-        """
-        async with self.config.guild(ctx.guild).raffles() as r:
-
-            raffle_data = r.get(raffle, None)
-
-            badges_list = raffle_data.get("badges_needed_to_enter", None)
-
-            if badges_list is None:
-                return await ctx.send(_("There are no required badges."))
-
-            del raffle_data["badges_needed_to_enter"]
-            message = _("Required bages list cleared for this raffle.")
-
-    @edit.group()
-    async def allowed(self, ctx):
-        """Manage the allowed users list in a raffle."""
-        pass
-
-    @allowed.command(name="add")
-    async def allowed_add(self, ctx, raffle: RaffleFactoryConverter, member: discord.Member):
-        """Add a member to the allowed list of a raffle.
-
-        **Arguments:**
-            - `<raffle>` - The name of the raffle.
-            - `<member>` - The member to add to the allowed list.
-        """
-        async with self.config.guild(ctx.guild).raffles() as r:
-
-            raffle_data = r.get(raffle, None)
-
-            allowed = raffle_data.get("allowed_users", [])
-
-            if member.id in allowed:
-                return await ctx.send(_("This user is already allowed in this raffle."))
-
-            if not allowed:
-                raffle_data["allowed_users"] = [member.id]
-            else:
-                allowed.append(member.id)
-
-            await ctx.send(_("{} added to the allowed list for this raffle.".format(member.name)))
-
-        await self.clean_guild_raffles(ctx)
-
-    @allowed.command(name="remove", aliases=["del"])
-    async def allowed_remove(self, ctx, raffle: RaffleFactoryConverter, member: discord.Member):
-        """Remove a member from the allowed list of a raffle.
-
-        **Arguments:**
-            - `<raffle>` - The name of the raffle.
-            - `<member>` - The member to remove from the allowed list.
-        """
-        async with self.config.guild(ctx.guild).raffles() as r:
-
-            raffle_data = r.get(raffle, None)
-
-            allowed = raffle_data.get("allowed_users", [])
-
-            if member.id not in allowed:
-                return await ctx.send(_("This user was not already allowed in this raffle."))
-
-            allowed.remove(member.id)
-            if not allowed:
-                del raffle_data["allowed_users"]
-            await ctx.send(
-                _("{} removed from the allowed list for this raffle.".format(member.name))
-            )
-
-        await self.clean_guild_raffles(ctx)
-
-    @allowed.command(name="clear")
-    async def allowed_clear(self, ctx, raffle: RaffleFactoryConverter):
-        """Clear the allowed list for a raffle."""
-        async with self.config.guild(ctx.guild).raffles() as r:
-
-            raffle_data = r.get(raffle, None)
-
-            allowed = raffle_data.get("allowed_users", None)
-
-            if allowed is None:
-                return await ctx.send(_("There are no allowed users."))
-
-        message = _("Are you sure you want to clear the allowed list for this raffle?")
-        can_react = ctx.channel.permissions_for(ctx.me).add_reactions
-        if not can_react:
-            message += " (y/n)"
-        message = await ctx.send(message)
-        if can_react:
-            start_adding_reactions(message, ReactionPredicate.YES_OR_NO_EMOJIS)
-            predicate = ReactionPredicate.yes_or_no(message, ctx.author)
-            event_type = "reaction_add"
-        else:
-            predicate = MessagePredicate.yes_or_no(ctx)
-            event_type = "message"
-
-        try:
-            await self.bot.wait_for(event_type, check=predicate, timeout=30)
-        except asyncio.TimeoutError:
-            await ctx.send(_("You took too long to respond."))
-            return
-
-        if predicate.result:
-            with contextlib.suppress(KeyError):
-                # Still wanna remove empty list here
-                del raffle_data["allowed_users"]
-            try:
-                await message.edit(content=_("Allowed list cleared for this raffle."))
-            except discord.NotFound:
-                await ctx.send(_("Allowed list cleared for this raffle."))
-
-        else:
-            await ctx.send(_("No changes have been made."))
-
-        await self.clean_guild_raffles(ctx)
-
-    @edit.group()
-    async def rolesreq(self, ctx):
-        """Manage role requirements in a raffle."""
-        pass
-
-    @rolesreq.command(name="add")
-    async def rolesreq_add(self, ctx, raffle: RaffleFactoryConverter, role: discord.Role):
-        """Add a role to the role requirements list of a raffle.
-
-        **Arguments:**
-            - `<raffle>` - The name of the raffle.
-            - `<role>` - The role to add to the list of role requirements.
-        """
-        async with self.config.guild(ctx.guild).raffles() as r:
-
-            raffle_data = r.get(raffle, None)
-
-            roles = raffle_data.get("roles_needed_to_enter", [])
-
-            if role.id in roles:
-                return await ctx.send(_("This role is already a requirement in this raffle."))
-
-            if not roles:
-                raffle_data["roles_needed_to_enter"] = [role.id]
-            else:
-                roles.append(role.id)
-
-            await ctx.send(
-                _("{} added to the role requirement list for this raffle.".format(role.name))
-            )
-
-        await self.clean_guild_raffles(ctx)
-
-    @rolesreq.command(name="remove", aliases=["del"])
-    async def rolereq_remove(self, ctx, raffle: RaffleFactoryConverter, role: discord.Role):
-        """Remove a role from the role requirements list of a raffle.
-
-        **Arguments:**
-            - `<raffle>` - The name of the raffle.
-            - `<role>` - The role to remove from the list of role requirements.
-        """
-        async with self.config.guild(ctx.guild).raffles() as r:
-
-            raffle_data = r.get(raffle, None)
-
-            roles = raffle_data.get("roles_needed_to_enter", [])
-
-            if role.id not in roles:
-                return await ctx.send(_("This role is not already a requirement in this raffle."))
-
-            roles.remove(role.id)
-            await ctx.send(
-                _("{} remove from the role requirement list for this raffle.".format(role.name))
-            )
-
-        await self.clean_guild_raffles(ctx)
-
-    @rolesreq.command(name="clear")
-    async def rolereq_clear(self, ctx, raffle: RaffleFactoryConverter):
-        """Clear the role requirement list for a raffle.
-
-
-        **Arguments:**
-            - `<raffle>` - The name of the raffle.
-        """
-        async with self.config.guild(ctx.guild).raffles() as r:
-
-            raffle_data = r.get(raffle, None)
-
-            rolesreq = raffle_data.get("roles_needed_to_enter", [])
-
-            if rolesreq is None:
-                return await ctx.send(_("There are no required roles."))
-
-            message = _(
-                "Are you sure you want to clear the role requirement list for this raffle?"
-            )
-            can_react = ctx.channel.permissions_for(ctx.me).add_reactions
-            if not can_react:
-                message += " (yes/no)"
-            message = await ctx.send(message)
-            if can_react:
-                start_adding_reactions(message, ReactionPredicate.YES_OR_NO_EMOJIS)
-                predicate = ReactionPredicate.yes_or_no(message, ctx.author)
-                event_type = "reaction_add"
-            else:
-                predicate = MessagePredicate.yes_or_no(ctx)
-                event_type = "message"
-
-            try:
-                await self.bot.wait_for(event_type, check=predicate, timeout=30)
-            except asyncio.TimeoutError:
-                await ctx.send(_("You took too long to respond."))
-                return
-
-            if predicate.result:
-                with contextlib.suppress(KeyError):
-                    # Still wanna remove empty list here
-                    del raffle_data["roles_needed_to_enter"]
-                message = "Role requirement list cleared for this raffle."
-                try:
-                    await message.edit(content=message)
-                except discord.NotFound:
-                    await ctx.send(message)
-
-            else:
-                await ctx.send(_("No changes have been made."))
-
-            await self.clean_guild_raffles(ctx)
