--- raffle/commands/management/locales/messages.pot
@@ -1,116 +1,0 @@
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: PACKAGE VERSION\n"
-"POT-Creation-Date: 2021-06-20 12:01+0000\n"
-"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
-"Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
-"Language-Team: LANGUAGE <LL@li.org>\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Generated-By: redgettext 3.3\n"
-
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/events.py:44
-msgid "There are no participants yet for this raffle."
-msgstr ""
-
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/events.py:52
-msgid "Congratulations {winner.mention}, you have won the {raffle} raffle!"
-msgstr ""
-
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/events.py:58
-msgid "Picking a winner from the pool..."
-msgstr ""
-
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/events.py:92
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/events.py:209
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/events.py:237
-msgid "You are not the owner of this raffle."
-msgstr ""
-
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/events.py:95
-msgid "This user has not entered this raffle."
-msgstr ""
-
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/events.py:98
-msgid "User removed from the raffle."
-msgstr ""
-
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/events.py:120
-msgid "You are already in this raffle."
-msgstr ""
-
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/events.py:124
-msgid "You are not allowed to join this particular raffle."
-msgstr ""
-
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/events.py:128
-msgid "You are not allowed to join this particular raffle"
-msgstr ""
-
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/events.py:132
-msgid "You cannot join your own raffle."
-msgstr ""
-
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/events.py:136
-msgid "Sorry, the maximum number of users have entered this raffle."
-msgstr ""
-
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/events.py:185
-msgid "You are not entered into this raffle."
-msgstr ""
-
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/events.py:212
-msgid "There are no entries yet for this raffle."
-msgstr ""
-
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/events.py:243
-msgid "Raffle ended."
-msgstr ""
-
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/misc.py:42
-msgid ""
-"Paste your YAML here. It will be validated, and if there is an exception, it"
-" will be returned to you."
-msgstr ""
-
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/misc.py:61
-msgid "This YAML is invalid."
-msgstr ""
-
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/misc.py:67
-msgid "An exception occured whilst parsing your data."
-msgstr ""
-
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/misc.py:70
-msgid "This YAML is good to go! No errors were found."
-msgstr ""
-
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/misc.py:81
-msgid "Raffles updated."
-msgstr ""
-
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/misc.py:83
-msgid "Everything was already up to date."
-msgstr ""
-
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/misc.py:93
-msgid "There are no ongoing raffles in this guild."
-msgstr ""
-
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/misc.py:96
-msgid "Are you sure you want to tear down all ongoing raffles in this guild?"
-msgstr ""
-
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/misc.py:112
-msgid "You took too long to respond."
-msgstr ""
-
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/misc.py:121
-msgid "Raffles cleared."
-msgstr ""
-
-#: /home/ubuntu/.local/share/Red-DiscordBot/data/kreusada/cogs/CogManager/cogs/raffle/commands/management/misc.py:124
-msgid "No changes have been made."
-msgstr ""
