+++ raffle/info.json
@@ -4,8 +4,8 @@
     ],
     "version": [
         1,
-        6,
-        12
+        7,
+        0
     ],
     "description": "[BETA] Create raffles for your guild, customizable with conditional blocks, and constructed using YAML.",
     "install_msg": "This cog is in pre-release so it is still being developed, but has been opened up for people to try out and test. Please report any issues or feature requests back to me in my support channel in the cog support server (`#support_kreusada-cogs`) https://discord.gg/GET4DVk\nThanks for installing, have fun. Please refer to my docs if you need any help: https://kreusadacogs.readthedocs.io/en/latest/cog_raffle.html",
